Refactor acceptance test code
